# blog-gin/docs/swagger.yaml
definitions:
...
  dto.CategoryBackDTO:
    properties:
      article_count:
        type: integer
      category_name:
        description: 分类名
        type: string
      created_at:
        description: 创建时间
        type: integer
      id:
        type: integer
      updated_at:
        description: 更新时间
        type: integer
    type: object
...
  dto.CategoryNewReq:
    properties:
      category_name:
        description: 分类名
        type: string
      id:
        type: integer
    type: object
  dto.CategoryQuery:
    properties:
      category_name:
        description: 分类名
        type: string
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      sorts:
        description: 排序
        items:
          type: string
        type: array
    type: object
...
  dto.FriendBackDTO:
    properties:
      created_at:
        description: 创建时间
        type: integer
      id:
        description: id
        type: integer
      link_address:
        description: 链接地址
        type: string
      link_avatar:
        description: 链接头像
        type: string
      link_intro:
        description: 链接介绍
        type: string
      link_name:
        description: 链接名
        type: string
      updated_at:
        description: 更新时间
        type: integer
    type: object
  dto.FriendNewReq:
    properties:
      id:
        description: id
        type: integer
      link_address:
        description: 链接地址
        type: string
      link_avatar:
        description: 链接头像
        type: string
      link_intro:
        description: 链接介绍
        type: string
      link_name:
        description: 链接名
        type: string
    type: object
  dto.FriendQuery:
    properties:
      link_name:
        description: 链接名
        type: string
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      sorts:
        description: 排序
        items:
          type: string
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.BatchResp:
    properties:
      success_count:
        type: integer
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdReq:
    properties:
      id:
        type: integer
    type: object
  github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.PageResp:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.FileBackDTO:
    properties:
      created_at:
        description: 创建时间
        type: integer
      file_md5:
        description: 文件md5值
        type: string
      file_name:
        description: 文件名称
        type: string
      file_path:
        description: 文件路径
        type: string
      file_size:
        description: 文件大小
        type: integer
      file_type:
        description: 文件类型
        type: string
      file_url:
        description: 上传路径
        type: string
      id:
        description: 文件目录ID
        type: integer
      updated_at:
        description: 更新时间
        type: integer
      user_id:
        description: 用户id
        type: string
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.MultiUploadFileReq:
    properties:
      file_path:
        description: 文件路径
        type: string
      files:
        description: 文件列表
        items: {}
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.UploadFileReq:
    properties:
      file:
        description: 文件
      file_path:
        description: 文件路径
        type: string
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support/
  description: APP接口鉴权方式：token采用自加密token的方式，自加密token规则：sha256(${x-timestamp}${base_secret})
    生成的加密token，x-timestamp 时间戳需要保持10分钟内
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger Example API
  version: 1.0.0
paths:
...
  /admin_api/v1/category/add_category:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryNewReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryBackDTO'
              type: object
      summary: '"创建文章分类"'
      tags:
      - Category
  /admin_api/v1/category/batch_delete_category:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.BatchResp'
              type: object
      summary: '"批量删除文章分类"'
      tags:
      - Category
  /admin_api/v1/category/delete_category:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.BatchResp'
              type: object
      summary: '"删除文章分类"'
      tags:
      - Category
  /admin_api/v1/category/find_category_list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryQuery'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.PageResp'
              type: object
      summary: '"分页获取文章分类列表"'
      tags:
      - Category
  /admin_api/v1/category/update_category:
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryNewReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryBackDTO'
              type: object
      summary: '"更新文章分类"'
      tags:
      - Category
...
  /admin_api/v1/friend/add_friend:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.FriendNewReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/dto.FriendBackDTO'
              type: object
      summary: '"创建友链"'
      tags:
      - Friend
  /admin_api/v1/friend/batch_delete_friend:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.BatchResp'
              type: object
      summary: '"批量删除友链"'
      tags:
      - Friend
  /admin_api/v1/friend/delete_friend:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.BatchResp'
              type: object
      summary: '"删除友链"'
      tags:
      - Friend
  /admin_api/v1/friend/find_friend_list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.FriendQuery'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_admin_dto.PageResp'
              type: object
      summary: '"分页获取友链列表"'
      tags:
      - Friend
  /admin_api/v1/friend/update_friend:
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.FriendNewReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/dto.FriendBackDTO'
              type: object
      summary: '"更新友链"'
      tags:
      - Friend
...
  /api/v1/file/multi_upload_file:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.MultiUploadFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.FileBackDTO'
                  type: array
              type: object
      summary: '"上传文件列表"'
      tags:
      - File
  /api/v1/file/upload_file:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.UploadFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_server_api_blog_dto.FileBackDTO'
              type: object
      summary: '"上传文件"'
      tags:
      - File
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"